Файл const.js переименован в constS.js. - Константы, используемые разово, перенесены из consts.js в те файлы, в которых они задействуются. - В файле route-presenter.js. цикл for заменен на цикл forEach.
