--- src/const.js
@@ -1,11 +1,0 @@
-const TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
-const CITIES = ['Berlin', 'Monaco', 'Barcelona', 'Kioto', 'Amsterdam', 'Hiroshima', 'Rome', 'Madrid', 'Tokio', 'Den Haag'];
-const LOREM_TEXT = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phaselus eros mauris, condime.';
-const MIN_BASE_PRICE = 300;
-const MAX_BASE_PRICE = 1100;
-const MINUTES_IN_DAY = 1440;
-const MINUTES_IN_HOUR = 60;
-const MAX_DAYS_BEFORE = 7;
-const MAX_DAYS_AFTER = 7;
-
-export {TYPES, CITIES, LOREM_TEXT, MIN_BASE_PRICE, MAX_BASE_PRICE, MINUTES_IN_DAY, MINUTES_IN_HOUR, MAX_DAYS_BEFORE, MAX_DAYS_AFTER};
