+++ src/consts.js
@@ -1,0 +1,3 @@
+const TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
+
+export {TYPES};
