+++ src/mock/destination.js
@@ -1,10 +1,11 @@
 import {getRandomInteger} from '../utils';
-import {CITIES, LOREM_TEXT} from '../const';
 
 //Функция для получения случайного названия города из массива cities
+const CITIES = ['Berlin', 'Monaco', 'Barcelona', 'Kioto', 'Amsterdam', 'Hiroshima', 'Rome', 'Madrid', 'Tokio', 'Den Haag'];
 const getCity = () => CITIES[getRandomInteger(0, CITIES.length - 1)];
 
 //Функции получения строки случайной длины для полей description и pictures.description
+const LOREM_TEXT = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phaselus eros mauris, condime.';
 const getRandomDescription = () => LOREM_TEXT.substring(0, (getRandomInteger(1, 6) * 50));
 const getRandomPictureDescription = () => LOREM_TEXT.substring(0, (getRandomInteger(3, 9) * 2));
